backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.11.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai==1.84.0
numpy==2.2.0
PyPDF2==3.0.1
python-multipart==0.0.20
